make lib file for modify functions
