+++ apps/newsletters-ui/src/app/components/edition-layouts/LayoutEditor.tsx
@@ -1,7 +1,8 @@
 import { Box, Button, Card, Divider, Stack } from "@mui/material";
-import { Layout, LayoutGroup, NewsletterData } from "@newsletters-nx/newsletters-data-client";
+import { Layout, NewsletterData } from "@newsletters-nx/newsletters-data-client";
 import { Fragment, useState } from "react";
 import { fetchPostApiData } from "../../api-requests/fetch-api-data";
+import { addNewGroup, deleteGroup, deleteNewsletterFromGroup, insertNewsletterIntoGroup, updateLayoutGroup } from "../../lib/modify-layout";
 import { StringInput } from "../SchemaForm/StringInput";
 import { NewsletterCard } from "./NewsletterCard";
 
@@ -9,76 +10,6 @@ interface Props {
     layout: Layout;
     editionId: string;
     newsletters: NewsletterData[];
-}
-
-const makeBlankGroup = () => ({
-    title: '',
-    newsletters: []
-})
-
-const updateLayoutGroup = (layout: Layout, groupIndex: number, groupMod: Partial<LayoutGroup>): Layout => {
-    const groupToUpdate = layout.groups[groupIndex];
-    if (!groupToUpdate) {
-        console.warn('invalid groupIndex', groupIndex, layout)
-        return layout
-    }
-    return {
-        ...layout,
-        groups: [
-            ...layout.groups.slice(0, groupIndex),
-            { ...groupToUpdate, ...groupMod },
-            ...layout.groups.slice(groupIndex + 1)
-        ]
-    }
-}
-
-const deleteNewsletterFromGroup = (layout: Layout, groupIndex: number, newsletterIndex: number): Layout => {
-    const groupToUpdate = layout.groups[groupIndex];
-    if (!groupToUpdate) {
-        console.warn('invalid groupIndex', groupIndex, layout)
-        return layout
-    }
-    return updateLayoutGroup(layout, groupIndex, {
-        newsletters: [
-            ...groupToUpdate.newsletters.slice(0, newsletterIndex),
-            ...groupToUpdate.newsletters.slice(newsletterIndex + 1),
-        ]
-    })
-}
-const insertNewsletterIntoGroup = (layout: Layout, groupIndex: number, newsletterIndex: number, newsletterId: string): Layout => {
-    const groupToUpdate = layout.groups[groupIndex];
-    if (!groupToUpdate) {
-        console.warn('invalid groupIndex', groupIndex, layout)
-        return layout
-    }
-    return updateLayoutGroup(layout, groupIndex, {
-        newsletters: [
-            ...groupToUpdate.newsletters.slice(0, newsletterIndex),
-            newsletterId,
-            ...groupToUpdate.newsletters.slice(newsletterIndex),
-        ]
-    })
-}
-
-const addNewGroup = (layout: Layout, groupIndex: number): Layout => {
-    return {
-        ...layout,
-        groups: [
-            ...layout.groups.slice(0, groupIndex),
-            makeBlankGroup(),
-            ...layout.groups.slice(groupIndex),
-        ]
-    }
-}
-
-const deleteGroup = (layout: Layout, groupIndexToInsertBefore: number): Layout => {
-    return {
-        ...layout,
-        groups: [
-            ...layout.groups.slice(0, groupIndexToInsertBefore),
-            ...layout.groups.slice(groupIndexToInsertBefore + 1),
-        ]
-    }
 }
